Day One
I added and created the game menu, i ama able to move between all the scenes( game, controls, how to play and quite the game).

I have also created the character controler, which allows the player to  move and look arouns aswell as jump. the player can also shoot.

Created a basic enemy character which moves around waypoints randomly  om a nav mesh on the plane,

i added health for the enemy and when the player shoots his weapon the player loses lives and then dies when all the haelth is lost.(need to hook it up the the health bar above the enemy)