Day Two
I added a scence fader into the game, for example when the player starts the game it will fade in and then when th player presses the game level button on the menu. it will fade out and fade back in to the game level.

i fixed the enemy health bar so it decreases when the enmy takes damage in the game, when the enemies health is 0 the enemy dies and then deleates itself from the game.

i also wrote a script so the player can gain health from the power up that will be in the game.

last thing i did was model the players base in blender and added it into unity.