Day Three
Fixed the players Health, for some reason when ever the player would walk into an object that the player would lose health from, the player would not lose health.

Added the building blocks and grid scripts from the percedural generating unity game. So that they can be used to create the city blocks.

Created a script so that the enemies can spawn after a certain amount of time, and  between waves there is a time limit of 15 seconds.

The enmies spawn but after that the enemies do not move towards the players base on the nav mesh/