Day Four
Finished the players gun in blender and exported it as an fbx into unity and attached it to the player character. I also removed the temporary player weapon from the game.

Added the procedural  generation grid to the plane.

I have written a script to spawn the buildings into the game when the play button is pressed. At the moment I have one stack of buildings spawning on the plane, and I will edit it so that they will spawn all over the map in city blocks. so that the player can run around the city blocks.