Day Five
Have basic city spawning when the play button is pressed. need to work more on it and model more builds so it is different. the city spawns using perlin noise into city blocks.

player shooting stoped working, i was using raycasting to damage the enemies, but now i am instantsiating the bullets from a fire point at the top of the gun.