30/04/2020
Parented the players gun to the enemy.

When the player get within a certtain distance the enmy will chase the player. If the player gets away then the enemy will stop chasing.

If the player gets within the enemies shooting range then the enemy will shoot at the player. If the player escapes then the enemy will stop shooting the player.

When the enemies lose all there health, they destroy themselves.