02/02/2020
Fixed the health of the enemies. when the enemies took damage from the player, all the enemies in the scene took damage. Now only the enemy that gets hit with the bullet will take damage.

Fixed the enemy movement.  now when the enemies spawns in, they will move towards the player, when the player gets into the look radius.