05/05/2020
Fixed the enemies movement. now the enemies move towards the players base when they spawn in.

Fixed the builds spawning in with regards to the nav mesh. Now the nav mesh updates at run time  allowing the base building blocks to carve out the nav mesh so that the enemies mive around them.