10/05/2020
Fixed the base health. Now when a enemy collides with the base the base will lose health. When the health reaches 0 game is over.

added the enemy model into the game, with sound effects for walking and for when the enemies shoot.

added a credits scene so when the player wins or loses the game the credits will be played.