25/05/2020
Worked on the enemie making sure that It plays well,.

I fixed the problem with the health as I had  the health as (if healh is == 0 ) instead of (if health is <= 0). This was causing the enemy to be kept alive if the health of the enmy was below 0.

Added the explosion particle effect to the cars as they are destroyed.

Added static walls around the map, as I was unable to get the was spawning with the perlin noise.

I added colliders to all the building parts, so that the player would not jump through the buildings.