26/05/2020
I created a new enmy that will explode when the enemy gets to clode to the player.

I worked on the UI to make it look nicer.

I fixed the bullet damage of the enmeies and the player.

I recorded the video and uploaded it to youtube and embeded it into the read me file of the project.